+++ bittorrent/src/main/java/org/johnnei/javatorrent/network/socket/TcpSocket.java
@@ -1,4 +1,4 @@
-package org.johnnei.javatorrent.internal.network.socket;
+package org.johnnei.javatorrent.network.socket;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -7,10 +7,16 @@ import java.net.InetSocketAddress;
 import java.net.Socket;
 import java.util.Objects;
 
+/**
+ * A Socket implementation which utilizes TCP to connect the two endpoints.
+ */
 public class TcpSocket implements ISocket {
 
 	private Socket socket;
 
+	/**
+	 * Creates a new unconnected socket.
+	 */
 	public TcpSocket() {
 		socket = new Socket();
 	}